save park data to json file
