+++ controllers/parkCtrl.js
@@ -1,11 +1,10 @@
 var fs = require('fs');
 var express = require('express');
 var request = require('request');
-var rp = require('request-promise-native');
+// var rp = require('request-promise-native');
 require('dotenv').config();
 const parkAPI = process.env.PARK_API;
 var bearerToken = process.env.TWITTER_BEARER_TOKEN; //the bearer token obtained from the last script
-let natParks = []; // array for storing park objects
 
 // gets information of all parks
 module.exports.getParks = (req, res, next) => {
@@ -32,36 +31,36 @@ module.exports.getParks = (req, res, next) => {
 };
 
 // gets parks from NPS API for initial storing to the database
-let extractParks = (req, res, next) => {
-  request.get(`https://developer.nps.gov/api/v1/parks?api_key=${parkAPI}&limit=519`, (err, response, body) => {
-    if (!err && response.statusCode == 200) {
-      var parks = JSON.parse(body).data;
-      // makes list of only the results that are National Parks
-      var allParks = parks.filter((park) => {
-        return park.designation==="National Park"
-      })
-      for(var i = 0; i < allParks.length; i++) { //loops over each park oject for inserting into Park table
-        const { Park } = req.app.get('models');
-        let savePark = {
-          fullName: allParks[i].fullName,
-          name: allParks[i].name,
-          parkCode: allParks[i].parkCode,
-          description: allParks[i].description,
-          states: allParks[i].states,
-          weatherInfo: allParks[i].weatherInfo,
-          url: allParks[i].url
-        }
-        Park.create(savePark)
-        .then( () => {
-          module.exports.getParks(req, res, next)
-        })
-        .catch( (err) => {
-          next(err);
-        }); 
-      }
-    }
-  })
-}
+// let extractParks = (req, res, next) => {
+//   request.get(`https://developer.nps.gov/api/v1/parks?api_key=${parkAPI}&limit=519`, (err, response, body) => {
+//     if (!err && response.statusCode == 200) {
+//       var parks = JSON.parse(body).data;
+//       // makes list of only the results that are National Parks
+//       var allParks = parks.filter((park) => {
+//         return park.designation==="National Park"
+//       })
+//       for(var i = 0; i < allParks.length; i++) { //loops over each park oject for inserting into Park table
+//         const { Park } = req.app.get('models');
+//         let savePark = {
+//           fullName: allParks[i].fullName,
+//           name: allParks[i].name,
+//           parkCode: allParks[i].parkCode,
+//           description: allParks[i].description,
+//           states: allParks[i].states,
+//           weatherInfo: allParks[i].weatherInfo,
+//           url: allParks[i].url
+//         }
+//         Park.create(savePark)
+//         .then( () => {
+//           module.exports.getParks(req, res, next)
+//         })
+//         .catch( (err) => {
+//           next(err);
+//         }); 
+//       }
+//     }
+//   })
+// }
 
 let onePark =[]
 module.exports.getSinglePark = (req, res, next) => {
