+++ package.json
@@ -5,6 +5,7 @@
 		"scripts": {
 				"start": "nodemon app.js",
 				"test": "echo \"Error: no test specified\" && exit 1",
+				"populate-parks": "node populate-parks.js",
 				"build-css": "./bin/build-css",
 				"watch-css": "./bin/watch-css"
 		},
