+++ routes/index.js
@@ -3,9 +3,9 @@
 const { Router } = require('express');
 const router = Router();
 
-const {
-  getParks
-} = require('../controllers/parkCtrl.js');
+// const {
+//   getParks
+// } = require('../controllers/parkCtrl.js');
 
 router.get('/', (req, res, next) => {
   res.render('index');
