+++ routes/parksRoute.js
@@ -9,6 +9,7 @@ const {
   savePark
 } = require('../controllers/parkCtrl.js');
 
+
 router.get('/parks', getParks); // gets list of all park names
 router.get('/parks/:id', getSinglePark); // details of a single park
 router.post('/parks/:id/:fullName', isLoggedIn, savePark, getSinglePark); // adds park to favorites table
